# app/Modules/Profile/Http/Requests/VerifyEmailRequest.php
<?php

namespace App\Modules\Profile\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

class VerifyEmailRequest extends FormRequest
{
    public function authorize(): bool
    {
        return true;
    }

    public function rules(): array
    {
        return [
            'code' => ['required', 'string', 'digits:'.config('moncolis.otp.length')],
        ];
    }
}