# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "MonColis Particulier API",
        "version": "0.1.0",
        "description": "API backend de **MonColis Particulier** (app mobile Flutter).\r\n\r\n- Base : `/api` \u2014 JSON en camelCase, r\u00e9ponses envelopp\u00e9es dans `{ \"data\": ... }`.\r\n- Authentification : `POST /login_check` (ou `/register`) renvoie un `token` \u00e0 passer en `Authorization: Bearer {token}`.\r\n- En-t\u00eate `client-app-code` : obligatoire lorsque le serveur a des codes configur\u00e9s.\r\n- Erreurs m\u00e9tier : `{ \"message\": \"...\", \"code\": \"INVALID_CREDENTIALS\" }` \u2014 le `code` est stable, l'app peut s'y fier.\r\n- Erreurs de validation : `422 { \"message\": \"...\", \"errors\": { \"champ\": [\"...\"] } }`.\r\n\r\nSwagger UI : [/docs/swagger](/docs/swagger) \u2014 Document OpenAPI brut : [/docs/api.json](/docs/api.json)."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
...
    "security": [
        {
            "bearerAuth": [],
            "clientAppCode": []
        }
    ],
    "paths": {
        "/auth/otp/request": {
            "post": {
                "operationId": "auth.otp.request",
                "description": "Envoie un code \u00e0 6 chiffres par SMS. M\u00eame r\u00e9ponse que le num\u00e9ro ait d\u00e9j\u00e0 un compte ou non.\n`devCode` n'est renvoy\u00e9 qu'en environnement local.\n\nErreur m\u00e9tier : `429 OTP_COOLDOWN` (nouvel envoi avant `resendIn` secondes).",
                "summary": "Recevoir un code de connexion",
                "tags": [
                    "Connexion par SMS"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhoneCodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "channel": {
                                                    "type": "string",
                                                    "const": "sms"
                                                },
                                                "destination": {
                                                    "type": "string"
                                                },
                                                "expiresIn": {
                                                    "type": "integer"
                                                },
                                                "resendIn": {
                                                    "type": "integer"
                                                },
                                                "devCode": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "channel",
                                                "destination",
                                                "expiresIn",
                                                "resendIn"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": [
                    {
                        "clientAppCode": []
                    }
                ]
            }
        },
...
        "/profile/password": {
            "put": {
                "operationId": "accountSecurity.updatePassword",
                "description": "Par d\u00e9faut (`logoutOtherDevices`), les autres appareils sont d\u00e9connect\u00e9s ; le jeton courant reste valide.\n`currentPassword` n'est pas demand\u00e9 si le compte n'a pas encore de mot de passe (`hasPassword: false`,\ncompte cr\u00e9\u00e9 par code SMS) : c'est alors la cr\u00e9ation du premier mot de passe.",
                "summary": "Changer mon mot de passe",
                "tags": [
                    "S\u00e9curit\u00e9 du compte"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/profile/phone/otp": {
            "post": {
                "operationId": "accountSecurity.requestPhoneChange",
                "description": "Envoie un code \u00e0 6 chiffres par SMS au nouveau num\u00e9ro. `devCode` n'est renvoy\u00e9 qu'en environnement local.\n\nErreurs m\u00e9tier : `429 OTP_COOLDOWN` (nouvel envoi avant `resendIn` secondes), `422 SAME_PHONE`.",
                "summary": "Changer de num\u00e9ro \u2014 1. envoyer le code",
                "tags": [
                    "S\u00e9curit\u00e9 du compte"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestPhoneChangeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "destination": {
                                                    "type": "string"
                                                },
                                                "expiresIn": {
                                                    "type": "integer"
                                                },
                                                "resendIn": {
                                                    "type": "integer"
                                                },
                                                "devCode": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "destination",
                                                "expiresIn",
                                                "resendIn"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/profile/phone/verify": {
            "post": {
                "operationId": "accountSecurity.confirmPhoneChange",
                "description": "Erreurs m\u00e9tier : `422 OTP_INVALID`, `422 OTP_EXPIRED`, `429 OTP_TOO_MANY_ATTEMPTS` (5 essais).",
                "summary": "Changer de num\u00e9ro \u2014 2. valider le code",
                "tags": [
                    "S\u00e9curit\u00e9 du compte"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmPhoneChangeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProfileResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ProfileResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "preferences"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/profile/email/otp": {
            "post": {
                "operationId": "accountSecurity.requestEmailVerification",
                "description": "Envoie un code \u00e0 6 chiffres \u00e0 l'adresse e-mail du profil.\n\nErreurs m\u00e9tier : `422 EMAIL_MISSING`, `422 EMAIL_ALREADY_VERIFIED`, `429 OTP_COOLDOWN`.",
                "summary": "V\u00e9rifier mon e-mail \u2014 1. envoyer le code",
                "tags": [
                    "S\u00e9curit\u00e9 du compte"
                ],
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "channel": {
                                                    "type": "string",
                                                    "const": "email"
                                                },
                                                "destination": {
                                                    "type": "string"
                                                },
                                                "expiresIn": {
                                                    "type": "integer"
                                                },
                                                "resendIn": {
                                                    "type": "integer"
                                                },
                                                "devCode": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "channel",
                                                "destination",
                                                "expiresIn",
                                                "resendIn"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/profile/email/verify": {
            "post": {
                "operationId": "accountSecurity.confirmEmailVerification",
                "description": "Erreurs m\u00e9tier : `422 OTP_INVALID`, `422 OTP_EXPIRED`, `429 OTP_TOO_MANY_ATTEMPTS`.",
                "summary": "V\u00e9rifier mon e-mail \u2014 2. valider le code",
                "tags": [
                    "S\u00e9curit\u00e9 du compte"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyEmailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProfileResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ProfileResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "preferences"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Jeton renvoy\u00e9 par /login_check ou /register.",
                "scheme": "bearer"
            },
            "clientAppCode": {
                "type": "apiKey",
                "description": "Code de l'application cliente (v\u00e9rifi\u00e9 si MONCOLIS_CLIENT_APP_CODES est renseign\u00e9).",
                "in": "header",
                "name": "client-app-code"
            }
        },
        "schemas": {
            "ConfirmPhoneChangeRequest": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                },
                "required": [
                    "phone",
                    "code"
                ],
                "title": "ConfirmPhoneChangeRequest"
            },
...
            "PhoneCodeRequest": {
                "type": "object",
                "description": "Step 1 of the SMS sign-in flow: the phone number.",
                "properties": {
                    "phone": {
                        "type": "string"
                    }
                },
                "required": [
                    "phone"
                ],
                "title": "PhoneCodeRequest"
            },
            "PreferencesResource": {
                "type": "object",
                "properties": {
                    "language": {
                        "type": "string"
                    },
                    "theme": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "notifyPush": {
                        "type": "boolean"
                    },
                    "notifySms": {
                        "type": "boolean"
                    },
                    "notifyEmail": {
                        "type": "boolean"
                    },
                    "marketingOptIn": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "language",
                    "theme",
                    "notifyPush",
                    "notifySms",
                    "notifyEmail",
                    "marketingOptIn"
                ],
                "title": "PreferencesResource"
            },
            "ProfileResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "firstName": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "lastName": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "fullName": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "phoneVerified": {
                        "type": "boolean"
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "emailVerified": {
                        "type": "boolean"
                    },
                    "hasPassword": {
                        "type": "boolean",
                        "description": "false for accounts created by SMS code until a password is set."
                    },
                    "avatarUrl": {
                        "type": "null"
                    },
                    "gender": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "birthDate": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "countryCode": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "profile": {
                        "type": "string",
                        "description": "Legacy field read by the Flutter UserModel.",
                        "const": "PARTICULIER"
                    },
                    "memberSince": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "completion": {
                        "type": "object",
                        "properties": {
                            "percent": {
                                "type": "integer"
                            },
                            "missing": {
                                "type": "array",
                                "items": {}
                            }
                        },
                        "required": [
                            "percent",
                            "missing"
                        ]
                    },
                    "preferences": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/PreferencesResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "required": [
                    "id",
                    "username",
                    "firstName",
                    "lastName",
                    "fullName",
                    "displayName",
                    "phone",
                    "phoneVerified",
                    "email",
                    "emailVerified",
                    "hasPassword",
                    "avatarUrl",
                    "gender",
                    "birthDate",
                    "countryCode",
                    "profile",
                    "memberSince",
                    "completion"
                ],
                "title": "ProfileResource"
            },
...
            "RequestPhoneChangeRequest": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    }
                },
                "required": [
                    "phone"
                ],
                "title": "RequestPhoneChangeRequest"
            },
...
            "UpdatePasswordRequest": {
                "type": "object",
                "properties": {
                    "currentPassword": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Accounts created by SMS code have no password yet: the first one is simply set."
                    },
                    "newPassword": {
                        "type": "string"
                    },
                    "newPasswordConfirmation": {
                        "type": "string"
                    },
                    "logoutOtherDevices": {
                        "type": "boolean",
                        "description": "Yango-like \"log out other devices\" (default: yes)."
                    }
                },
                "required": [
                    "newPassword",
                    "newPasswordConfirmation"
                ],
                "title": "UpdatePasswordRequest"
            },
...
            "VerifyEmailRequest": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    }
                },
                "required": [
                    "code"
                ],
                "title": "VerifyEmailRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}